Add files via upload
An OTP (One-Time Password) generator is a security feature designed to enhance the authentication process by adding an additional layer of protection beyond traditional usernames and passwords. The purpose of OTPs is to provide a dynamic and time-sensitive code that users must input during the login process. This helps mitigate the risks associated with static passwords, such as unauthorized access due to theft, hacking, or phishing.